Add FIB2 service and test case (#336)
* Update host vector (WIP)

+++ types/accumulatecontext.go
@@ -21,7 +21,7 @@ type Kai_state struct {
 	Kai_m uint32            `json:"chi_m"` // The index of the bless service
 	Kai_a uint32            `json:"chi_a"` // The index of the designate service
 	Kai_v uint32            `json:"chi_v"` // The index of the assign service
-	Kai_g map[uint32]uint32 `json:"chi_g"` // g is a small dictionary containing the indices of services which automatically accumulate in each block together with a basic amount of gas with which each accumulates
+	Kai_g map[uint32]uint64 `json:"chi_g"` // g is a small dictionary containing the indices of services which automatically accumulate in each block together with a basic amount of gas with which each accumulates
 }
 
 // fixed size for the authorization queue
